pyproject.toml
[tool.poetry]
name = "breakableminoreditor"
version = "0.1.0"
description = ""
authors = ["repl.it user <replituser@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
wordcloud = "^1.8.1"
Pillow = "^8.0.1"
matplotlib = "^3.3.3"
numpy = "^1.19.4"
pandas = "^1.2.0"
bs4 = "^0.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
